[fix] correção de tipos

+++ include/pacman.h
@@ -9,9 +9,9 @@
 // Atributos do Pacman
 typedef struct pacman {
     coord_t position;               // coordenada cartesiana
-    int radius;                     // raio de visão
-    int count_mov;                  // contador de passos
-    int pellets;                    // contador de pastilhas coletadas
+    uint32_t radius;                // raio de visão
+    uint32_t count_mov;             // contador de passos
+    uint16_t pellets;               // contador de pastilhas coletadas
 } pacman_t;
 
 // Move o Pacman na direção desejada se for uma posição válida no mapa (não-parede)
